refactor av MapPage component in i mindre komponenter og gjenbrukbar logikk i egne filer
