+++ src/utils/mapStyleConfig.ts
@@ -1,0 +1,660 @@
+import type { FeatureCollection } from "geojson";
+import type { Hospital, PoliceStation } from "../services/MapDataService";
+
+/**
+ * Creates the map style configuration with all sources and layers
+ * @returns The complete map style configuration object
+ */
+export const createMapStyle = (options: {
+  geoJSONData: FeatureCollection;
+  policeStations: PoliceStation[];
+  hospitals: Hospital[];
+  showPoliceStations: boolean;
+  showHospitals: boolean;
+  showRoads: boolean;
+}): mapboxgl.Style => {
+  const {
+    geoJSONData,
+    policeStations,
+    hospitals,
+    showPoliceStations,
+    showHospitals,
+    showRoads,
+  } = options;
+
+  return {
+    version: 8,
+    sources: {
+      "vector-tiles": {
+        type: "vector",
+        url: "https://api.maptiler.com/tiles/v3/tiles.json?key=eE87Cs6ofbIAP2G5mFFy",
+      },
+      "roads-wms": {
+        type: "raster",
+        tiles: [
+          "https://wms.geonorge.no/skwms1/wms.vegnett2?SERVICE=WMS&REQUEST=GetMap&VERSION=1.3.0&FORMAT=image/png&TRANSPARENT=true&LAYERS=Vegnett2&WIDTH=256&HEIGHT=256&CRS=EPSG:3857&STYLES=&BBOX={bbox-epsg-3857}",
+        ],
+        tileSize: 256,
+        attribution: "&copy; Geonorge - Vegnett",
+      },
+      "geojson-source": {
+        type: "geojson",
+        data: geoJSONData || { type: "FeatureCollection", features: [] },
+      },
+      "police-source": {
+        type: "geojson",
+        data:
+          policeStations && policeStations.length > 0
+            ? {
+                type: "FeatureCollection",
+                features: policeStations.map((station) => ({
+                  type: "Feature",
+                  geometry: {
+                    type: "Point",
+                    coordinates: [station.lon, station.lat],
+                  },
+                  properties: {
+                    id: station.id,
+                    name: station.name,
+                    phone: station.phone,
+                  },
+                })),
+              }
+            : { type: "FeatureCollection", features: [] },
+      },
+      "hospital-source": {
+        type: "geojson",
+        data:
+          hospitals && hospitals.length > 0
+            ? {
+                type: "FeatureCollection",
+                features: hospitals.map((hospital) => ({
+                  type: "Feature",
+                  geometry: {
+                    type: "Point",
+                    coordinates: [hospital.lon, hospital.lat],
+                  },
+                  properties: {
+                    id: hospital.id,
+                    name: hospital.name,
+                    phone: hospital.phone,
+                  },
+                })),
+              }
+            : { type: "FeatureCollection", features: [] },
+      },
+    },
+    layers: [
+      {
+        id: "background",
+        type: "background",
+        paint: {
+          "background-color": "#1a1a1a",
+        },
+      },
+      {
+        id: "water",
+        type: "fill",
+        source: "vector-tiles",
+        "source-layer": "water",
+        paint: {
+          "fill-color": "#222222",
+        },
+      },
+      {
+        id: "landcover",
+        type: "fill",
+        source: "vector-tiles",
+        "source-layer": "landcover",
+        paint: {
+          "fill-color": "#2a2a2a",
+        },
+      },
+      {
+        id: "roads",
+        type: "line",
+        source: "vector-tiles",
+        "source-layer": "transportation",
+        paint: {
+          "line-color": "#404040",
+          "line-width": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            10,
+            1,
+            15,
+            2,
+            20,
+            4,
+          ],
+        },
+      },
+      {
+        id: "roads-wms-layer",
+        type: "raster",
+        source: "roads-wms",
+        paint: {
+          "raster-opacity": showRoads ? 1 : 0,
+        },
+      },
+      {
+        id: "shelter-heatmap",
+        type: "heatmap",
+        source: "geojson-source",
+        maxzoom: 11,
+        paint: {
+          "heatmap-weight": [
+            "interpolate",
+            ["linear"],
+            ["get", "population"],
+            0,
+            0,
+            1000,
+            0.5,
+            5000,
+            1,
+          ],
+          "heatmap-intensity": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            0,
+            1,
+            11,
+            3,
+          ],
+          "heatmap-color": [
+            "interpolate",
+            ["linear"],
+            ["heatmap-density"],
+            0,
+            "rgba(0, 0, 255, 0)",
+            0.2,
+            "rgba(0, 0, 255, 0.5)",
+            0.4,
+            "rgba(0, 255, 255, 0.7)",
+            0.6,
+            "rgba(255, 255, 0, 0.8)",
+            0.8,
+            "rgba(255, 128, 0, 0.9)",
+            1,
+            "rgba(255, 0, 0, 1)",
+          ],
+          "heatmap-radius": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            0,
+            4,
+            11,
+            30,
+          ],
+          "heatmap-opacity": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            10,
+            1,
+            11,
+            0,
+          ],
+        },
+      },
+      {
+        id: "geojson-layer-heat",
+        type: "circle",
+        source: "geojson-source",
+        minzoom: 11,
+        paint: {
+          "circle-radius": [
+            "interpolate",
+            ["exponential", 1.75],
+            ["zoom"],
+            12,
+            30,
+            14,
+            45,
+            16,
+            60,
+          ],
+          "circle-color": "#ffc400",
+          "circle-opacity": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            12,
+            0.2,
+            14,
+            0.15,
+            15,
+            0.1,
+          ],
+          "circle-blur": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            12,
+            1,
+            14,
+            0.8,
+            16,
+            0.6,
+          ],
+        },
+      },
+      {
+        id: "geojson-layer-glow",
+        type: "circle",
+        source: "geojson-source",
+        minzoom: 11,
+        paint: {
+          "circle-radius": [
+            "interpolate",
+            ["exponential", 1.75],
+            ["zoom"],
+            12,
+            20,
+            14,
+            30,
+            16,
+            40,
+          ],
+          "circle-color": "#ffcd38",
+          "circle-opacity": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            12,
+            0.4,
+            14,
+            0.3,
+            15,
+            0.2,
+          ],
+          "circle-blur": 1,
+        },
+      },
+      {
+        id: "geojson-layer",
+        type: "circle",
+        source: "geojson-source",
+        minzoom: 11,
+        paint: {
+          "circle-radius": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            12,
+            6,
+            14,
+            12,
+            16,
+            14,
+          ],
+          "circle-color": "#ffc400",
+          "circle-stroke-width": 2,
+          "circle-stroke-color": "#FFFFFF",
+          "circle-opacity": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            12,
+            0.9,
+            14,
+            1,
+          ],
+        },
+      },
+      {
+        id: "geojson-layer-symbols",
+        type: "symbol",
+        source: "geojson-source",
+        minzoom: 11,
+        layout: {
+          "icon-image": "biohazard-icon",
+          "icon-size": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            12,
+            0,
+            13,
+            0.4,
+            14,
+            0.5,
+            16,
+            0.6,
+          ],
+          "icon-allow-overlap": true,
+          "icon-ignore-placement": true,
+          "icon-anchor": "center",
+          "icon-offset": [0, 0],
+        },
+        paint: {
+          "icon-opacity": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            12,
+            0,
+            13,
+            0.7,
+            14,
+            1,
+          ],
+        },
+      },
+      // Police layers
+      {
+        id: "police-layer-heat",
+        type: "circle",
+        source: "police-source",
+        minzoom: 11,
+        layout: {
+          visibility: showPoliceStations ? "visible" : "none",
+        },
+        paint: {
+          "circle-radius": [
+            "interpolate",
+            ["exponential", 1.75],
+            ["zoom"],
+            12,
+            30,
+            14,
+            45,
+            16,
+            60,
+          ],
+          "circle-color": "#0066ff",
+          "circle-opacity": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            12,
+            0.2,
+            14,
+            0.15,
+            15,
+            0.1,
+          ],
+          "circle-blur": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            12,
+            1,
+            14,
+            0.8,
+            16,
+            0.6,
+          ],
+        },
+      },
+      {
+        id: "police-layer-glow",
+        type: "circle",
+        source: "police-source",
+        minzoom: 11,
+        layout: {
+          visibility: showPoliceStations ? "visible" : "none",
+        },
+        paint: {
+          "circle-radius": [
+            "interpolate",
+            ["exponential", 1.75],
+            ["zoom"],
+            12,
+            20,
+            14,
+            30,
+            16,
+            40,
+          ],
+          "circle-color": "#3884ff",
+          "circle-opacity": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            12,
+            0.4,
+            14,
+            0.3,
+            15,
+            0.2,
+          ],
+          "circle-blur": 1,
+        },
+      },
+      {
+        id: "police-layer",
+        type: "circle",
+        source: "police-source",
+        minzoom: 11,
+        layout: {
+          visibility: showPoliceStations ? "visible" : "none",
+        },
+        paint: {
+          "circle-radius": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            12,
+            6,
+            14,
+            12,
+            16,
+            14,
+          ],
+          "circle-color": "#0066ff",
+          "circle-stroke-width": 2,
+          "circle-stroke-color": "#FFFFFF",
+          "circle-opacity": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            12,
+            0.9,
+            14,
+            1,
+          ],
+        },
+      },
+      {
+        id: "police-layer-symbols",
+        type: "symbol",
+        source: "police-source",
+        minzoom: 11,
+        layout: {
+          visibility: showPoliceStations ? "visible" : "none",
+          "icon-image": "handcuffs-icon",
+          "icon-size": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            12,
+            0,
+            13,
+            0.4,
+            14,
+            0.5,
+            16,
+            0.6,
+          ],
+          "icon-allow-overlap": true,
+          "icon-ignore-placement": true,
+          "icon-anchor": "center",
+          "icon-offset": [-3, -2],
+        },
+        paint: {
+          "icon-opacity": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            12,
+            0,
+            13,
+            0.7,
+            14,
+            1,
+          ],
+        },
+      },
+      // Hospital layers
+      {
+        id: "hospital-layer-heat",
+        type: "circle",
+        source: "hospital-source",
+        minzoom: 11,
+        layout: {
+          visibility: showHospitals ? "visible" : "none",
+        },
+        paint: {
+          "circle-radius": [
+            "interpolate",
+            ["exponential", 1.75],
+            ["zoom"],
+            12,
+            30,
+            14,
+            45,
+            16,
+            60,
+          ],
+          "circle-color": "#ff0000",
+          "circle-opacity": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            12,
+            0.2,
+            14,
+            0.15,
+            15,
+            0.1,
+          ],
+          "circle-blur": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            12,
+            1,
+            14,
+            0.8,
+            16,
+            0.6,
+          ],
+        },
+      },
+      {
+        id: "hospital-layer-glow",
+        type: "circle",
+        source: "hospital-source",
+        minzoom: 11,
+        layout: {
+          visibility: showHospitals ? "visible" : "none",
+        },
+        paint: {
+          "circle-radius": [
+            "interpolate",
+            ["exponential", 1.75],
+            ["zoom"],
+            12,
+            20,
+            14,
+            30,
+            16,
+            40,
+          ],
+          "circle-color": "#ff3333",
+          "circle-opacity": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            12,
+            0.4,
+            14,
+            0.3,
+            15,
+            0.2,
+          ],
+          "circle-blur": 1,
+        },
+      },
+      {
+        id: "hospital-layer",
+        type: "circle",
+        source: "hospital-source",
+        minzoom: 11,
+        layout: {
+          visibility: showHospitals ? "visible" : "none",
+        },
+        paint: {
+          "circle-radius": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            12,
+            6,
+            14,
+            12,
+            16,
+            14,
+          ],
+          "circle-color": "#ff0000",
+          "circle-stroke-width": 2,
+          "circle-stroke-color": "#FFFFFF",
+          "circle-opacity": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            12,
+            0.9,
+            14,
+            1,
+          ],
+        },
+      },
+      {
+        id: "hospital-layer-symbols",
+        type: "symbol",
+        source: "hospital-source",
+        minzoom: 11,
+        layout: {
+          visibility: showHospitals ? "visible" : "none",
+          "icon-image": "hospital-icon",
+          "icon-size": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            12,
+            0,
+            13,
+            0.4,
+            14,
+            0.5,
+            16,
+            0.6,
+          ],
+          "icon-allow-overlap": true,
+          "icon-ignore-placement": true,
+          "icon-anchor": "center",
+          "icon-offset": [-3, -2],
+        },
+        paint: {
+          "icon-opacity": [
+            "interpolate",
+            ["linear"],
+            ["zoom"],
+            12,
+            0,
+            13,
+            0.7,
+            14,
+            1,
+          ],
+        },
+      },
+    ],
+  };
+};
